VE: split the files based on type of audits
